Applied a patch supplied by Leonardo Vitor da Silva that updates ExEnAndroid to work with the latest version of Mono for Android.
(Merged internal revision 162)

git-svn-id: https://exen.svn.codeplex.com/svn@15692 7a4bfd9b-e295-4cbd-a51f-ab472bcf69b7

+++ ExEnAndroid/Audio/SoundEffect.cs
@@ -11,7 +11,7 @@ namespace Microsoft.Xna.Framework.Audio
 		internal static SoundPool pool;
 		internal static void Setup()
 		{
-			pool = new Android.Media.SoundPool(32, (int)Android.Media.Stream.Music, 0);
+			pool = new Android.Media.SoundPool(32, Android.Media.Stream.Music, 0);
 		}
 	
 		#endregion
